split between what is part of the proposal and what is a real implementation not related to OT API
