+++ ext/README.md
@@ -1,0 +1,5 @@
+# External modules
+
+This Python package includes some extensions that are not
+part of the proposal. All modules here are only used to provide
+examples or simplified implementations, for a proof of concept.
